@@ -1,16 +1,16 @@
 package com.spicymemes.core.network
 
-import net.minecraft.network.PacketBuffer
-import net.minecraftforge.fml.network.NetworkEvent
+import net.minecraft.network.*
+import net.minecraftforge.fmllegacy.network.*
 import java.util.function.Supplier
 
 interface SpicyPacketHandler<MSG : SpicyPacket> {
 
     fun process(packet: MSG, ctx: NetworkEvent.Context)
 
-    fun encode(packet: MSG, buf: PacketBuffer)
+    fun encode(packet: MSG, buf: FriendlyByteBuf)
 
-    fun decode(buf: PacketBuffer): MSG
+    fun decode(buf: FriendlyByteBuf): MSG
 
     fun handle(message: MSG, ctxSupplier: Supplier<NetworkEvent.Context>) {
         ctxSupplier.get().also { ctx ->
